+++ silence/db/dal.py
@@ -2,6 +2,7 @@ from pymysql.cursors import DictCursor
 
 from silence.db.connector import get_conn
 from silence.exceptions import DatabaseError
+from silence.decorators import db_call
 
 ###############################################################################
 # The DAL (Data Access Layer) functions provide an abstraction layer
@@ -53,3 +54,16 @@ def update(q, params=None):
     finally:
         # Close the cursor
         cursor.close()
+
+###############################################################################
+# Safe wrappers for the API, which return an HTTPError instead of a
+# DatabaseError
+###############################################################################
+
+@db_call
+def api_safe_query(*args, **kwargs):
+    return query(*args, **kwargs)
+
+@db_call
+def api_safe_update(*args, **kwargs):
+    return update(*args, **kwargs)