+++ silence/server/endpoint_loader.py
@@ -3,6 +3,9 @@ import importlib
 from os import listdir
 from os.path import splitext
 
+from silence.server import manager, default_endpoints
+from silence.settings import settings
+
 ###############################################################################
 # Look for .py files inside the project's "/api" folder
 # and force them to run, activating the @endpoint decorators
@@ -18,3 +21,13 @@ def load_user_endpoints():
         except ImportError:
             raise RuntimeError(f"Could not load the API file {module_name}")
     
+###############################################################################
+# Register the Silence-provided endpoints
+###############################################################################
+def load_default_endpoints():
+    route_prefix = settings.API_PREFIX
+    if route_prefix.endswith("/"):
+        route_prefix = route_prefix[:-1]
+
+    manager.APP.add_url_rule(f"{route_prefix}/login", "login", default_endpoints.login, methods=["POST"])
+    manager.APP.add_url_rule(f"{route_prefix}/register", "register", default_endpoints.register, methods=["POST"])