+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 0.3.0
+- Added support for the default endpoints /login and /register using custom tables and fields
+- Added support for session tokens and authentication requirements
+
 # 0.2.0
 - Added support for HTTP POST/PUT/DELETE and SQL INSERT/UPDATE/DELETE
 - Added support for static web files